+++ CHANGELOG.md
@@ -6,6 +6,17 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [0.0.7]
+
+- Tables sorting (even advanced multi-column scenarios are supported)
+- Table filtering (both value based and rule based supported).
+- Kusto query sections can now be hidden / viewed just like URL sections.
+- Red squiggly line support for Kusto editor.
+- Auto-complete for Kusto editor is now context aware, multi-line aware, variable aware and their suggestions are expected to match the red squiggly line behavior.
+- New ability to add cluster + database connection pairs as favorites and pick them.
+- Fix: Switching favorites across clusters no longer reuses the wrong cluster client.
+
+
 ## [0.0.6]
 
 - Bug fixes and UX improvements